[FIX] Fixed plot_point_brain() function
Currently, in the plot_point_brain() function, the xyz aspect ratios are manually adjusted to [0.57, 0.57, 0.60]. This manual adjustement provides nice-looking brains that are unfortunately not accurate representations of the actual coordinates of the brain. The way the brain will look will also be totally dependent on the values of the coordinates that we input. This might not be something that we want. I therefore changed the way we adjust the aspect ratio of the different axes such that the ratio is accurately depicting the relationship between the different axes.